+++ D20Tek.Spectre.Console.Extensions.UnitTests/Mocks/MockAsyncCommand.cs
@@ -12,7 +12,7 @@ internal class MockAsyncCommand(IAnsiConsole console) : AsyncCommand
 {
     private readonly IAnsiConsole _writer = console;
 
-    public override Task<int> ExecuteAsync(CommandContext context, CancellationToken cancellation)
+    protected override Task<int> ExecuteAsync(CommandContext context, CancellationToken cancellation)
     {
         _writer.WriteLine("Success!");
         return Task.FromResult(0);