+++ D20Tek.Spectre.Console.Extensions.UnitTests/Mocks/MockAsyncCommandWithException.cs
@@ -10,7 +10,7 @@ namespace D20Tek.Spectre.Console.Extensions.UnitTests.Mocks;
 
 internal class MockAsyncCommandWithException : AsyncCommand
 {
-    public override Task<int> ExecuteAsync(CommandContext context, CancellationToken cancellation)
+    protected override Task<int> ExecuteAsync(CommandContext context, CancellationToken cancellation)
     {
         throw new ArgumentOutOfRangeException(nameof(context));
     }